Refactoring objects, things, resources, etc. into a new submodule called "core".
