+++ zombpyg/core/object.py
