+++ zombpyg/core/things.py
@@ -1,6 +1,24 @@
 import random
-import pygame
 import numpy
+import pygame
+
+from zombpyg.core import Thing, CircularThing, RectangularThing, FightingThing
+from zombpyg.sensor import Sensor
+from zombpyg.utils.surroundings import Color, get_movement_estimates
+from zombpyg.utils.geometry import _valid_angle, get_angle_and_distance_to_point
+from zombpyg.weapons import ZombieClaws, Knife, Axe, Gun, Rifle, Shotgun
+from zombpyg.actions import (
+    MoveableThing,
+    RotatableThing,
+    AttackingThing,
+    ExecutableAction,
+    ForwardMoveAction,
+    RightMoveAction,
+    BackwardMoveAction,
+    LeftMoveAction,
+    RotateAction,
+    AttackAction
+)
 
 
 class Thing(object):
@@ -129,3 +147,26 @@ class FightingThing(CircularThing):
         if y >= world.h: y = world.h - 1
 
         return x, y
+
+# TODO: Remove the following for now
+class Box(RectangularThing):
+    """Solid box."""
+    MAX_LIFE = 10
+
+    def __init__(self, left, top, width, height):
+        super(Box, self).__init__(
+            left, top, width, height,
+            u'box', 'yellow', Box.MAX_LIFE
+        )
+
+class DeadBody(CircularThing):
+    """Dead body."""
+    def __init__(self, r, name, color, display_time=2.0):
+        super(DeadBody, self).__init__(
+            0, 0, r,
+            name, color, 0,
+        )
+        self.life = display_time
+
+    def decrement_life(self, t):
+        self.life -= t
